Move StatusOr to core/platform
PiperOrigin-RevId: 367726383
